refactor: Move threat model prompt builder into stride_gpt.core.prompts
create_threat_model_prompt and its two OWASP section helpers
(create_llm_stride_prompt_section, create_agentic_stride_prompt_section)
move from threat_model.py into stride_gpt/core/prompts.py, where the
rest of the prompt construction already lives. cli.py drops the
sys.path.insert + top-level "from threat_model import ..." workaround
and imports directly from the package. threat_model.py keeps a thin
re-export shim (with __all__) so the legacy Streamlit tabs and any
external callers continue to work unchanged.
